+++ spec/features/universe_graphic_spec.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe 'Get graphic information' do
+  before { VCR.insert_cassette 'esi/universe/graphic/20481' }
+
+  after { VCR.eject_cassette }
+
+  let(:options) { { id: 20_481 } }
+
+  subject { EveOnline::ESI::UniverseGraphic.new(options) }
+
+  specify { expect(subject.scope).to eq(nil) }
+
+  specify do
+    expect(subject.as_json).to eq(collision_file: nil,
+                                  graphic_file: nil,
+                                  graphic_id: 20_481,
+                                  icon_folder: nil,
+                                  sof_dna: 'ai1_t1:tash-murkon:amarr',
+                                  sof_fation_name: 'tash-murkon',
+                                  sof_hull_name: 'ai1_t1',
+                                  sof_race_name: 'amarr')
+  end
+end