+++ CHANGELOG.md
@@ -9,6 +9,12 @@
 * Add `EveOnline::ESI::UniverseGroups`
 * Add `EveOnline::ESI::Models::Group`
 * Add `EveOnline::ESI::UniverseGroup`
+* Add `EveOnline::ESI::Models::Graphic`
+* Add `EveOnline::ESI::UniverseGraphic`
+* Add `EveOnline::ESI::UniverseSystem`
+* Add `EveOnline::ESI::Models::System`
+* Add `EveOnline::ESI::Models::Planet`
+* Add `EveOnline::ESI::Models::Planets`
 
 ## v0.19.0